+++ CopilotKit/packages/runtime/src/lib/runtime/remote-action-constructors.ts
@@ -23,7 +23,8 @@ import { execute } from "./remote-lg-action";
 import { CopilotKitError, CopilotKitLowLevelError } from "@copilotkit/shared";
 import { writeJsonLineResponseToEventStream } from "../streaming";
 import { CopilotKitApiDiscoveryError, ResolvedCopilotKitError } from "@copilotkit/shared";
-import { parseJson } from "@copilotkit/shared";
+import { parseJson, tryMap } from "@copilotkit/shared";
+import { ActionInput } from "../../graphql/inputs/action.input";
 
 export function constructLGCRemoteAction({
   endpoint,
@@ -82,10 +83,10 @@ export function constructLGCRemoteAction({
           state,
           configurable,
           properties: graphqlContext.properties,
-          actions: actionInputsWithoutAgents.map((action) => ({
+          actions: tryMap(actionInputsWithoutAgents, (action: ActionInput) => ({
             name: action.name,
             description: action.description,
-            parameters: parseJson(action.jsonSchema, "") as string,
+            parameters: JSON.parse(action.jsonSchema),
           })),
           metaEvents,
         });
@@ -232,10 +233,10 @@ export function constructRemoteActions({
                 state,
                 configurable,
                 properties: graphqlContext.properties,
-                actions: actionInputsWithoutAgents.map((action) => ({
+                actions: tryMap(actionInputsWithoutAgents, (action: ActionInput) => ({
                   name: action.name,
                   description: action.description,
-                  parameters: parseJson(action.jsonSchema, {}),
+                  parameters: JSON.parse(action.jsonSchema),
                 })),
                 metaEvents,
               }),