+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,27 @@
 # ui
 
+## 1.6.0-next.6
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.6.0-next.6
+- @copilotkit/shared@1.6.0-next.6
+
+## 1.6.0-next.5
+
+### Patch Changes
+
+- Updated dependencies [090203d]
+  - @copilotkit/shared@1.6.0-next.5
+  - @copilotkit/runtime-client-gql@1.6.0-next.5
+
+## 1.6.0-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.6.0-next.4
+- @copilotkit/shared@1.6.0-next.4
+
 ## 1.6.0-next.3
 
 ### Patch Changes