+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @copilotkit/runtime-client-gql
 
+## 1.6.0-next.6
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.6
+
+## 1.6.0-next.5
+
+### Patch Changes
+
+- Updated dependencies [090203d]
+  - @copilotkit/shared@1.6.0-next.5
+
+## 1.6.0-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.4
+
 ## 1.6.0-next.3
 
 ### Patch Changes