+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @copilotkit/shared
 
+## 1.6.0-next.6
+
+## 1.6.0-next.5
+
+### Patch Changes
+
+- 090203d: - fix: use tryMap method to filter out possibly invalid items
+
+## 1.6.0-next.4
+
 ## 1.6.0-next.3
 
 ## 1.6.0-next.2